refactor: seperate auth and user controller
