+++ backend/src/main/java/com/example/backend/domain/auth/AuthController.java
@@ -1,0 +1,67 @@
+package com.example.backend.domain.auth;
+
+import com.example.backend.domain.auth.dto.*;
+import com.example.backend.domain.user.Role;
+import com.example.backend.global.security.JwtTokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/auth")
+@RequiredArgsConstructor
+public class AuthController {
+    private final AuthService authService;
+    private final RefreshTokenService refreshTokenService;
+    private final JwtTokenProvider jwtTokenProvider;
+
+    @PostMapping("/sign-up")
+    public ResponseEntity<String> signup(@RequestBody SignupRequest signupRequest) {
+        authService.signup(signupRequest);
+        return ResponseEntity.status(HttpStatus.CREATED).body("회원가입 성공");
+    }
+
+    @PostMapping("/sign-in")
+    public ResponseEntity<TokenResponse> login(@RequestBody LoginRequest loginRequest) {
+        TokenResponse tokens = authService.login(loginRequest);
+        return ResponseEntity.ok(tokens);
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(@RequestBody TokenRequest request){
+        String refreshToken = request.getRefreshToken();
+
+        if (!jwtTokenProvider.validateToken(refreshToken)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String userId = jwtTokenProvider.getUserIdFromToken(refreshToken);
+
+        refreshTokenService.deleteRefreshToken(userId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/refresh")
+    public ResponseEntity<TokenResponse> refreshToken(@RequestBody TokenRequest request){
+        String refreshToken = request.getRefreshToken();
+
+        if(!jwtTokenProvider.validateToken(refreshToken)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String userId= jwtTokenProvider.getUserIdFromToken(refreshToken);
+        String role = jwtTokenProvider.getRoleFromToken(refreshToken);
+
+        String savedRefreshToken = refreshTokenService.getRefreshToken(userId);
+        if (savedRefreshToken == null || !savedRefreshToken.equals(refreshToken)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String newAccessToken = jwtTokenProvider.generateAccessToken(userId, Role.valueOf(role));
+        String newRefreshToken = jwtTokenProvider.generateRefreshToken(userId, Role.valueOf(role));
+
+        return ResponseEntity.ok(new TokenResponse(newAccessToken, newRefreshToken));
+    }
+
+}
