+++ backend/src/main/java/com/example/backend/domain/user/UserController.java
@@ -1,47 +1,19 @@
 package com.example.backend.domain.user;
 
-import com.example.backend.domain.user.dto.*;
-import com.example.backend.global.security.JwtTokenProvider;
+import com.example.backend.domain.user.dto.RewardRequest;
+import com.example.backend.domain.user.dto.UserInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/auth")
+@RequestMapping("/user")
 @RequiredArgsConstructor
 public class UserController {
+
     private final UserService userService;
-    private final RefreshTokenService refreshTokenService;
-    private final JwtTokenProvider jwtTokenProvider;
-
-    @PostMapping("/sign-up")
-    public ResponseEntity<String> signup(@RequestBody SignupRequest signupRequest) {
-        userService.signup(signupRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).body("회원가입 성공");
-    }
-
-    @PostMapping("/sign-in")
-    public ResponseEntity<TokenResponse> login(@RequestBody LoginRequest loginRequest) {
-        TokenResponse tokens = userService.login(loginRequest);
-        return ResponseEntity.ok(tokens);
-    }
-
-    @PostMapping("/logout")
-    public ResponseEntity<Void> logout(@RequestBody TokenRequest request){
-        String refreshToken = request.getRefreshToken();
-
-        if (!jwtTokenProvider.validateToken(refreshToken)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-
-        String userId = jwtTokenProvider.getUserIdFromToken(refreshToken);
-
-        refreshTokenService.deleteRefreshToken(userId);
-        return ResponseEntity.noContent().build();
-    }
 
     @GetMapping("/user-info")
     public ResponseEntity<UserInfo> getUserInfo(@AuthenticationPrincipal UserDetailsImpl userDetails){
@@ -60,28 +32,6 @@ public class UserController {
         );
 
         return ResponseEntity.ok(userInfo);
-    }
-
-    @PostMapping("/refresh")
-    public ResponseEntity<TokenResponse> refreshToken(@RequestBody TokenRequest request){
-        String refreshToken = request.getRefreshToken();
-
-        if(!jwtTokenProvider.validateToken(refreshToken)){
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-
-        String userId= jwtTokenProvider.getUserIdFromToken(refreshToken);
-        String role = jwtTokenProvider.getRoleFromToken(refreshToken);
-
-        String savedRefreshToken = refreshTokenService.getRefreshToken(userId);
-        if (savedRefreshToken == null || !savedRefreshToken.equals(refreshToken)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-
-        String newAccessToken = jwtTokenProvider.generateAccessToken(userId, Role.valueOf(role));
-        String newRefreshToken = jwtTokenProvider.generateRefreshToken(userId, Role.valueOf(role));
-
-        return ResponseEntity.ok(new TokenResponse(newAccessToken, newRefreshToken));
     }
 
     @PostMapping("/reward")
