+++ backend/src/main/java/com/example/backend/domain/user/UserRepository.java
@@ -2,8 +2,6 @@ package com.example.backend.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface UserRepository extends JpaRepository<User, String> {
 
 }
