+++ backend/src/main/java/com/example/backend/global/security/CustomUserDetailsService.java
@@ -4,13 +4,10 @@ import com.example.backend.domain.user.User;
 import com.example.backend.domain.user.UserDetailsImpl;
 import com.example.backend.domain.user.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
