+++ mobile/src/features/home/api/userApi.ts
@@ -1,0 +1,6 @@
+import api from "../../../shared/api";
+
+export async function getUserInfo(){
+  const response = await api.get('/user/user-info');
+  return response.data;
+}
