+++ mobile/src/features/listening/api/rewardApi.ts
@@ -2,7 +2,7 @@ import api from "../../../shared/api";
 
 export async function requestReward(coin: number) {
     try{
-        const res = await api.post('/auth/reward', {
+        const res = await api.post('/user/reward', {
             coin: coin
         });
         return res.data;
